ecommerce_web/requirements.txt
alembic==1.14.1
aniso8601==10.0.0
bcrypt==4.2.1
blinker==1.9.0
chardet==5.2.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.9
MarkupSafe==3.0.2
pillow==11.1.0
PyJWT==2.10.1
PyMySQL==1.1.1
python-dotenv==1.0.1
pytz==2025.1
reportlab==4.3.1
six==1.17.0
SQLAlchemy==2.0.38
typing_extensions==4.12.2
Werkzeug==3.1.3
